requirements.txt
Flask==2.2.2
Flask_Login==0.6.2
Flask-WTF==1.0.1
Markdown==3.4.1
passlib==1.7.4
requests==2.28.1
password_strength==0.0.3.post2
Flask-Limiter==3.1.0
bleach==5.0.1
pycryptodomex==3.16.0